Visitor pattern refactoring
