+++ GOF-design-patterns/GOF-design-patterns.csproj
@@ -43,6 +43,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BehavioralPatterns\Visitor.cs" />
     <Compile Include="StructuralPatterns\Bridge.cs" />
     <Compile Include="CreationalPatterns\AbstractFactory.cs" />
     <Compile Include="CreationalPatterns\Builder.cs" />
@@ -64,8 +65,6 @@
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="BehavioralPatterns\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
